extract styles from analysis screen into analysisStyles
